# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  opscore_backend_internal_document_interfaces_api_schema.CreateDocumentRequest:
    properties:
      access_scope:
        example: public
        type: string
      commit_hash:
        description: 'Optional: if empty, use latest commit'
        example: abc1234567890
        type: string
      file_path:
        example: docs/backup-procedure.md
        type: string
      is_auto_update:
        example: true
        type: boolean
      repository_id:
        example: a1b2c3d4-e5f6-7890-1234-567890abcdef
        type: string
    required:
    - access_scope
    - file_path
    - repository_id
    type: object
...
  opscore_backend_internal_document_interfaces_api_schema.DocumentResponse:
    properties:
      access_scope:
        example: public
        type: string
      created_at:
        example: "2025-04-22T10:00:00Z"
        type: string
      current_version:
        $ref: '#/definitions/opscore_backend_internal_document_interfaces_api_schema.DocumentVersionResponse'
      id:
        example: a1b2c3d4-e5f6-7890-1234-567890abcdef
        type: string
      is_auto_update:
        example: true
        type: boolean
      is_published:
        example: true
        type: boolean
      owner:
        example: database-team
        type: string
      repository_id:
        example: a1b2c3d4-e5f6-7890-1234-567890abcdef
        type: string
      updated_at:
        example: "2025-04-22T12:00:00Z"
        type: string
      version_count:
        example: 3
        type: integer
    type: object
...
  opscore_backend_internal_document_interfaces_api_schema.ErrorResponse:
    properties:
      code:
        example: VALIDATION_FAILED
        type: string
      details:
        additionalProperties: true
        type: object
      message:
        example: Validation failed
        type: string
    type: object
...
  opscore_backend_internal_document_interfaces_api_schema.ListDocumentsResponse:
    properties:
      documents:
        items:
          $ref: '#/definitions/opscore_backend_internal_document_interfaces_api_schema.DocumentListItemResponse'
        type: array
    type: object
...
  opscore_backend_internal_document_interfaces_api_schema.UpdateDocumentMetadataRequest:
    properties:
      access_scope:
        example: private
        type: string
      is_auto_update:
        example: false
        type: boolean
      owner:
        example: new-team
        type: string
    type: object
  opscore_backend_internal_document_interfaces_api_schema.UpdateDocumentRequest:
    properties:
      commit_hash:
        description: 'Optional: if empty, use latest commit'
        example: def4567890123
        type: string
      file_path:
        example: docs/backup-procedure.md
        type: string
    required:
    - file_path
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: This is the API documentation for the OpsCore backend service.
  title: OpsCore Backend API
  version: "1.0"
paths:
...
  /documents:
    get:
      description: Retrieves a list of all published documents
      parameters:
      - description: Filter by repository ID
        in: query
        name: repository_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved documents
          schema:
            $ref: '#/definitions/opscore_backend_internal_document_interfaces_api_schema.ListDocumentsResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/opscore_backend_internal_document_interfaces_api_schema.ErrorResponse'
      summary: List all documents
      tags:
      - documents
    post:
      consumes:
      - application/json
      description: Create a new document with an initial version
      parameters:
      - description: Document information
        in: body
        name: document
        required: true
        schema:
          $ref: '#/definitions/opscore_backend_internal_document_interfaces_api_schema.CreateDocumentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Document created successfully
          schema:
            $ref: '#/definitions/opscore_backend_internal_document_interfaces_api_schema.DocumentResponse'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/opscore_backend_internal_document_interfaces_api_schema.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/opscore_backend_internal_document_interfaces_api_schema.ErrorResponse'
      summary: Create a new document
      tags:
      - documents
  /documents/{docId}:
    get:
      description: Retrieves detailed information about a specific document by ID
      parameters:
      - description: Document ID
        in: path
        name: docId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved document details
          schema:
            $ref: '#/definitions/opscore_backend_internal_document_interfaces_api_schema.DocumentResponse'
        "400":
          description: Invalid document ID format
          schema:
            $ref: '#/definitions/opscore_backend_internal_document_interfaces_api_schema.ErrorResponse'
        "404":
          description: Document not found
          schema:
            $ref: '#/definitions/opscore_backend_internal_document_interfaces_api_schema.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/opscore_backend_internal_document_interfaces_api_schema.ErrorResponse'
      summary: Get document details
      tags:
      - documents
    put:
      consumes:
      - application/json
      description: Update an existing document by creating a new version
      parameters:
      - description: Document ID
        in: path
        name: docId
        required: true
        type: string
      - description: Document update information
        in: body
        name: document
        required: true
        schema:
          $ref: '#/definitions/opscore_backend_internal_document_interfaces_api_schema.UpdateDocumentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Document updated successfully
          schema:
            $ref: '#/definitions/opscore_backend_internal_document_interfaces_api_schema.DocumentResponse'
        "400":
          description: Invalid request body or document ID
          schema:
            $ref: '#/definitions/opscore_backend_internal_document_interfaces_api_schema.ErrorResponse'
        "404":
          description: Document not found
          schema:
            $ref: '#/definitions/opscore_backend_internal_document_interfaces_api_schema.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/opscore_backend_internal_document_interfaces_api_schema.ErrorResponse'
      summary: Update a document
      tags:
      - documents
  /documents/{docId}/metadata:
    patch:
      consumes:
      - application/json
      description: Updates document metadata such as owner, access scope, and auto-update
        setting
      parameters:
      - description: Document ID
        in: path
        name: docId
        required: true
        type: string
      - description: Metadata update information
        in: body
        name: metadata
        required: true
        schema:
          $ref: '#/definitions/opscore_backend_internal_document_interfaces_api_schema.UpdateDocumentMetadataRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Metadata updated successfully
          schema:
            $ref: '#/definitions/opscore_backend_internal_document_interfaces_api_schema.DocumentResponse'
        "400":
          description: Invalid request body or document ID
          schema:
            $ref: '#/definitions/opscore_backend_internal_document_interfaces_api_schema.ErrorResponse'
        "404":
          description: Document not found
          schema:
            $ref: '#/definitions/opscore_backend_internal_document_interfaces_api_schema.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/opscore_backend_internal_document_interfaces_api_schema.ErrorResponse'
      summary: Update document metadata
      tags:
      - documents
...
  /documents/{docId}/versions:
    get:
      description: Retrieves all versions for a specific document
      parameters:
      - description: Document ID
        in: path
        name: docId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved version history
          schema:
            $ref: '#/definitions/opscore_backend_internal_document_interfaces_api_schema.VersionHistoryResponse'
        "400":
          description: Invalid document ID format
          schema:
            $ref: '#/definitions/opscore_backend_internal_document_interfaces_api_schema.ErrorResponse'
        "404":
          description: Document not found
          schema:
            $ref: '#/definitions/opscore_backend_internal_document_interfaces_api_schema.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/opscore_backend_internal_document_interfaces_api_schema.ErrorResponse'
      summary: Get document version history
      tags:
      - documents
  /documents/{docId}/versions/{version}:
    get:
      description: Retrieves details of a specific version of a document
      parameters:
      - description: Document ID
        in: path
        name: docId
        required: true
        type: string
      - description: Version number
        in: path
        name: version
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved version details
          schema:
            $ref: '#/definitions/opscore_backend_internal_document_interfaces_api_schema.DocumentVersionResponse'
        "400":
          description: Invalid document ID or version number
          schema:
            $ref: '#/definitions/opscore_backend_internal_document_interfaces_api_schema.ErrorResponse'
        "404":
          description: Document or version not found
          schema:
            $ref: '#/definitions/opscore_backend_internal_document_interfaces_api_schema.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/opscore_backend_internal_document_interfaces_api_schema.ErrorResponse'
      summary: Get a specific document version
      tags:
      - documents
  /documents/{docId}/versions/{version}/publish:
    post:
      consumes:
      - application/json
      description: Publishes a specific version of a document, making it the current
        version
      parameters:
      - description: Document ID
        in: path
        name: docId
        required: true
        type: string
      - description: Version number
        in: path
        name: version
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Version published successfully
          schema:
            $ref: '#/definitions/opscore_backend_internal_document_interfaces_api_schema.DocumentResponse'
        "400":
          description: Invalid document ID or version number
          schema:
            $ref: '#/definitions/opscore_backend_internal_document_interfaces_api_schema.ErrorResponse'
        "404":
          description: Document or version not found
          schema:
            $ref: '#/definitions/opscore_backend_internal_document_interfaces_api_schema.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/opscore_backend_internal_document_interfaces_api_schema.ErrorResponse'
      summary: Publish a specific document version
      tags:
      - documents
  /documents/{docId}/versions/{version}/rollback:
    post:
      consumes:
      - application/json
      description: Rolls back the document to a previous version
      parameters:
      - description: Document ID
        in: path
        name: docId
        required: true
        type: string
      - description: Version number
        in: path
        name: version
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Rollback successful
          schema:
            $ref: '#/definitions/opscore_backend_internal_document_interfaces_api_schema.DocumentResponse'
        "400":
          description: Invalid document ID or version number
          schema:
            $ref: '#/definitions/opscore_backend_internal_document_interfaces_api_schema.ErrorResponse'
        "404":
          description: Document or version not found
          schema:
            $ref: '#/definitions/opscore_backend_internal_document_interfaces_api_schema.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/opscore_backend_internal_document_interfaces_api_schema.ErrorResponse'
      summary: Rollback to a previous document version
      tags:
      - documents
  /documents/publish:
    post:
      consumes:
      - application/json
      description: Publish a document by fetching it from a Git repository via OAuth
        connection
      parameters:
      - description: Document publish information
        in: body
        name: document
        required: true
        schema:
          $ref: '#/definitions/opscore_backend_internal_document_interfaces_api_schema.PublishDocumentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Document published successfully
          schema:
            $ref: '#/definitions/opscore_backend_internal_document_interfaces_api_schema.DocumentResponse'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/opscore_backend_internal_document_interfaces_api_schema.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/opscore_backend_internal_document_interfaces_api_schema.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/opscore_backend_internal_document_interfaces_api_schema.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Publish a document from an OAuth connection
      tags:
      - documents
...
securityDefinitions:
  BearerAuth:
    description: 'Enter "Bearer " followed by your JWT token. Example: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"